Move BuildRoot into the buildtool package
