+++ modules/core/src/main/scala/org/scalasteward/core/io/WorkspaceAlg.scala
@@ -20,7 +20,8 @@ import better.files.File
 import cats.FlatMap
 import cats.syntax.all._
 import org.scalasteward.core.application.Config
-import org.scalasteward.core.vcs.data.{BuildRoot, Repo}
+import org.scalasteward.core.buildtool.BuildRoot
+import org.scalasteward.core.vcs.data.Repo
 import org.typelevel.log4cats.Logger
 
 trait WorkspaceAlg[F[_]] {
