+++ modules/core/src/main/scala/org/scalasteward/core/scalafmt/ScalafmtAlg.scala
@@ -21,11 +21,12 @@ import cats.data.OptionT
 import cats.syntax.all._
 import io.circe.ParsingFailure
 import org.scalasteward.core.application.Config
+import org.scalasteward.core.buildtool.BuildRoot
 import org.scalasteward.core.data.{Scope, Version}
 import org.scalasteward.core.io.{FileAlg, ProcessAlg, WorkspaceAlg}
 import org.scalasteward.core.scalafmt.ScalafmtAlg.{opts, parseScalafmtConf}
 import org.scalasteward.core.util.Nel
-import org.scalasteward.core.vcs.data.{BuildRoot, Repo}
+import org.scalasteward.core.vcs.data.Repo
 import org.typelevel.log4cats.Logger
 
 final class ScalafmtAlg[F[_]](config: Config)(implicit
