+++ modules/core/src/test/scala/org/scalasteward/core/buildtool/mill/MillAlgTest.scala
@@ -2,12 +2,13 @@ package org.scalasteward.core.buildtool.mill
 
 import cats.effect.unsafe.implicits.global
 import munit.FunSuite
+import org.scalasteward.core.buildtool.BuildRoot
 import org.scalasteward.core.buildtool.mill.MillAlg.extractDeps
 import org.scalasteward.core.data.Version
 import org.scalasteward.core.mock.MockContext.context._
 import org.scalasteward.core.mock.MockState
 import org.scalasteward.core.mock.MockState.TraceEntry.Cmd
-import org.scalasteward.core.vcs.data.{BuildRoot, Repo}
+import org.scalasteward.core.vcs.data.Repo
 
 class MillAlgTest extends FunSuite {
   test("getDependencies") {
